% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult_income}
\alias{adult_income}
\title{Adult Income dataset (500-row sample)}
\format{
A tibble with 500 rows and 16 variables:
\describe{
\item{id}{Row identifier (integer)}
\item{age}{Age in years (integer)}
\item{workclass}{Employment type (character)}
\item{fnlwgt}{Final weight, a census sampling weight (integer)}
\item{education}{Highest level of education (character)}
\item{education_num}{Education encoded as an integer (integer)}
\item{marital_status}{Marital status (character)}
\item{occupation}{Occupation category (character)}
\item{relationship}{Relationship to householder (character)}
\item{race}{Race (character)}
\item{sex}{Sex (character)}
\item{capital_gain}{Capital gains (integer)}
\item{capital_loss}{Capital losses (integer)}
\item{hours_per_week}{Hours worked per week (integer)}
\item{native_country}{Country of origin (character)}
\item{income}{Income bracket: \verb{<=50K} or \verb{>50K} (character)}
}
}
\source{
\url{https://archive.ics.uci.edu/dataset/2/adult}
}
\usage{
adult_income
}
\description{
A 500-row random sample of the UCI Adult Income dataset, used in
package examples and vignettes.
}
\keyword{datasets}
