% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.rsdv_quality_report}
\alias{autoplot.rsdv_quality_report}
\title{Plot a quality report}
\usage{
\method{autoplot}{rsdv_quality_report}(object, ...)
}
\arguments{
\item{object}{An \code{rsdv_quality_report} object.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Produces a bar chart of per-column similarity scores, with a horizontal
line at the overall score.
}
