% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{check_constraint}
\alias{check_constraint}
\title{Check a single constraint against each row of a data frame}
\usage{
check_constraint(data, constraint)
}
\arguments{
\item{data}{A data frame.}

\item{constraint}{An \code{rsdv_constraint} object.}
}
\value{
Logical vector of length \code{nrow(data)}.
}
\description{
Check a single constraint against each row of a data frame
}
\examples{
df <- data.frame(a = c(1, 2, 3), b = c(1, 2, 9))
check_constraint(df, equality_constraint("a", "b"))
}
