% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{inequality_constraint}
\alias{inequality_constraint}
\title{Constraint: col_a must be less than / greater than col_b}
\usage{
inequality_constraint(col_a, col_b, type = c("lt", "lte", "gt", "gte"))
}
\arguments{
\item{col_a, col_b}{Column names (character).}

\item{type}{One of \code{"lt"}, \code{"lte"}, \code{"gt"}, \code{"gte"}.}
}
\value{
An \code{rsdv_constraint} object.
}
\description{
Constraint: col_a must be less than / greater than col_b
}
\examples{
inequality_constraint("low", "high", type = "lt")
}
