% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesizer.R
\name{is_fitted}
\alias{is_fitted}
\title{Check whether a synthesizer has been fitted}
\usage{
is_fitted(x)
}
\arguments{
\item{x}{A synthesizer object.}
}
\value{
\code{TRUE} if \code{\link[=fit]{fit()}} has been called; \code{FALSE} otherwise.
}
\description{
Check whether a synthesizer has been fitted
}
\examples{
\dontrun{
syn <- gaussian_copula_synthesizer(metadata())
is_fitted(syn)
}
}
