% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_io.R
\name{load_metadata}
\alias{load_metadata}
\title{Load metadata from a JSON file}
\usage{
load_metadata(path)
}
\arguments{
\item{path}{Path to a JSON file produced by \code{\link[=save_metadata]{save_metadata()}}.}
}
\value{
An \code{rsdv_metadata} object.
}
\description{
Load metadata from a JSON file
}
