% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{print.custom_constraint}
\alias{print.custom_constraint}
\title{Print method for a custom_constraint}
\usage{
\method{print}{custom_constraint}(x, ...)
}
\arguments{
\item{x}{A \code{custom_constraint} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for a custom_constraint
}
