% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_report.R
\name{print.rsdv_diagnostic_report}
\alias{print.rsdv_diagnostic_report}
\title{Print method for rsdv_diagnostic_report}
\usage{
\method{print}{rsdv_diagnostic_report}(x, ...)
}
\arguments{
\item{x}{An \code{rsdv_diagnostic_report} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for rsdv_diagnostic_report
}
