% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{print.rsdv_metadata}
\alias{print.rsdv_metadata}
\title{Print method for rsdv_metadata}
\usage{
\method{print}{rsdv_metadata}(x, ...)
}
\arguments{
\item{x}{An \code{rsdv_metadata} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for rsdv_metadata
}
\examples{
print(metadata())
}
