% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_report.R
\name{print.rsdv_privacy_report}
\alias{print.rsdv_privacy_report}
\title{Print method for rsdv_privacy_report}
\usage{
\method{print}{rsdv_privacy_report}(x, ...)
}
\arguments{
\item{x}{An \code{rsdv_privacy_report} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for rsdv_privacy_report
}
