% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_report.R
\name{print.rsdv_quality_report}
\alias{print.rsdv_quality_report}
\title{Print method for rsdv_quality_report}
\usage{
\method{print}{rsdv_quality_report}(x, ...)
}
\arguments{
\item{x}{An \code{rsdv_quality_report} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for rsdv_quality_report
}
