% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesizer.R
\name{validate_data}
\alias{validate_data}
\title{Validate that a data frame is compatible with metadata}
\usage{
validate_data(data, meta)
}
\arguments{
\item{data}{A data frame.}

\item{meta}{An \code{rsdv_metadata} object.}
}
\value{
Invisibly \code{TRUE}; throws an error if validation fails.
}
\description{
Checks that all columns registered in \code{meta} are present in \code{data}.
}
\examples{
\dontrun{
validate_data(adult_income, metadata(adult_income))
}
}
