% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.rsdv_privacy_report}
\alias{autoplot.rsdv_privacy_report}
\title{Plot a privacy report}
\usage{
\method{autoplot}{rsdv_privacy_report}(object, ...)
}
\arguments{
\item{object}{An \code{rsdv_privacy_report} object.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots the NNDR score as a gauge-style bar.
}
\examples{
\donttest{
syn <- gaussian_copula_synthesizer(metadata(adult_income)) |> fit(adult_income)
synth <- sample(syn, n = 500)
pr <- privacy_report(adult_income, synth)
ggplot2::autoplot(pr)
}
}
