% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{custom_constraint}
\alias{custom_constraint}
\title{Constraint: arbitrary row-wise predicate}
\usage{
custom_constraint(fn)
}
\arguments{
\item{fn}{A function \code{f(row)} accepting a one-row data frame, returning
a single logical.}
}
\value{
An \code{rsdv_constraint} object.
}
\description{
Constraint: arbitrary row-wise predicate
}
\examples{
custom_constraint(function(row) row$x > 0)
}
