% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{equality_constraint}
\alias{equality_constraint}
\title{Constraint: two columns must be equal row-wise}
\usage{
equality_constraint(col_a, col_b)
}
\arguments{
\item{col_a, col_b}{Column names (character).}
}
\value{
An \code{rsdv_constraint} object.
}
\description{
Constraint: two columns must be equal row-wise
}
\examples{
equality_constraint("city", "city_copy")
}
