% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{print.equality_constraint}
\alias{print.equality_constraint}
\title{Print method for an equality_constraint}
\usage{
\method{print}{equality_constraint}(x, ...)
}
\arguments{
\item{x}{An \code{equality_constraint} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for an equality_constraint
}
