% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{print.fixed_combinations_constraint}
\alias{print.fixed_combinations_constraint}
\title{Print method for a fixed_combinations_constraint}
\usage{
\method{print}{fixed_combinations_constraint}(x, ...)
}
\arguments{
\item{x}{A \code{fixed_combinations_constraint} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for a fixed_combinations_constraint
}
