% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{print.inequality_constraint}
\alias{print.inequality_constraint}
\title{Print method for an inequality_constraint}
\usage{
\method{print}{inequality_constraint}(x, ...)
}
\arguments{
\item{x}{An \code{inequality_constraint} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for an inequality_constraint
}
