% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_report.R
\name{print.rsdv_privacy_report}
\alias{print.rsdv_privacy_report}
\title{Print method for rsdv_privacy_report}
\usage{
\method{print}{rsdv_privacy_report}(x, ...)
}
\arguments{
\item{x}{An \code{rsdv_privacy_report} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for rsdv_privacy_report
}
\examples{
\dontrun{
syn <- gaussian_copula_synthesizer(metadata(adult_income)) |> fit(adult_income)
synth <- sample(syn, n = 500)
pr <- privacy_report(adult_income, synth)
print(pr)
}
}
